V0.2.3.4 Update uninstall_pm2.sh

+++ env_init/src/uninstall_pm2.sh
@@ -9,9 +9,13 @@ n='\033[0m'
 
 echo -e "${b}------------------- Pm2 卸载 -------------------${n}"
 
-. nvm use 23
-echo -e "${b}删除 pm2 工具${n}"
-npm uninstall pm2 -g
+if command -v nvm > /dev/null 2>&1; then
+    . nvm use 23
+    echo -e "${b}删除 pm2 工具${n}"
+    npm uninstall pm2 -g
+else
+    echo -e "${r}nvm 未安装，跳过 pm2 卸载${n}"
+fi
 sudo rm -rf ~/.pm2
 echo -e "${g}pm2 删除成功${n}"
